+++ changes.md
@@ -28,6 +28,8 @@
 * Added documentation for minibase module.
 * Updated greeble documentation.
 * Updated README.md
+* Added window documentation and examples.
+* Added damage documentation.
 
 ## 0.3.3
 * Upped cadqueryhelper version 0.2.1